week 5 task 1solutions
